Finish weapon statistics form. Add report helpers to begin formatting data.

+++ db/migrate/20180318214937_create_damage_reports.rb
@@ -1,9 +1,9 @@
 class CreateDamageReports < ActiveRecord::Migration[5.1]
   def change
     create_table :damage_reports do |t|
-      t.decimal :min_damage
-      t.decimal :max_damage
-      t.decimal :damage_per_second
+      t.float :min_damage
+      t.float :max_damage
+      t.float :damage_per_second
       t.references :weapon
       t.references :attack
