+++ db/migrate/20180318222704_create_attacks.rb
@@ -3,8 +3,8 @@ class CreateAttacks < ActiveRecord::Migration[5.1]
     create_table :attacks do |t|
       t.string :attack_name
       t.string :attack_type
-      t.decimal :APS_impact
-      t.decimal :DMG_impact
+      t.float :APS_impact
+      t.float :DMG_impact
       t.string :element
       t.references :character
